Refactor: change namespace pci to devices
