+++ common/lib/kernel/devices/pci_device.cpp
@@ -21,7 +21,7 @@
 #include <villas/kernel/devices/pci_device.hpp>
 #include <villas/utils.hpp>
 
-using namespace villas::kernel::pci;
+using namespace villas::kernel::devices;
 
 #define PCI_BASE_ADDRESS_N(n) (PCI_BASE_ADDRESS_0 + sizeof(uint32_t) * (n))
