+++ common/lib/kernel/vfio_group.cpp
@@ -68,7 +68,7 @@ std::shared_ptr<Device> Group::attachDevice(std::shared_ptr<Device> device) {
 
 std::shared_ptr<Device>
 Group::attachDevice(const std::string &name,
-                    const kernel::pci::PciDevice *pci_device) {
+                    const kernel::devices::PciDevice *pci_device) {
   auto device = std::make_shared<Device>(name, fd, pci_device);
   return attachDevice(device);
 }
