+++ fpga/include/villas/fpga/pcie_card.hpp
@@ -55,7 +55,7 @@ public:         // TODO: make this private
   bool doReset; // Reset VILLASfpga during startup?
   int affinity; // Affinity for MSI interrupts
 
-  std::shared_ptr<kernel::pci::PciDevice> pdev; // PCI device handle
+  std::shared_ptr<kernel::devices::PciDevice> pdev; // PCI device handle
 
 protected:
   Logger getLogger() const { return villas::Log::get(name); }
